% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis2dClass.R
\docType{class}
\name{basis2Dfd-class}
\alias{basis2Dfd-class}
\alias{basis2Dfd}
\alias{basis2Dfd,basisfd,basisfd-method}
\title{A Class of 2-dimensional functional data objects}
\usage{
\S4method{basis2Dfd}{basisfd,basisfd}(bs1, bs2)
}
\arguments{
\item{bs1}{a basisfd ...}

\item{bs2}{a basisfd ...}
}
\description{
The \code{basis2Dfd} class represents functional data ...
}
\section{Methods (by generic)}{
\itemize{
\item \code{basis2Dfd}: Constructor for basis2Dfd objects when
}}

\section{Slots}{

\describe{
\item{\code{bs1}}{a basisfd}

\item{\code{bs2}}{a basisfd ...}

\item{\code{support}}{a matrix....}
}}

\examples{
x <- 1
}
