% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis2dClass.R
\docType{methods}
\name{basis2Dfd-constructor}
\alias{basis2Dfd-constructor}
\title{Constructor for basis2Dfd objects}
\usage{
basis2Dfd(bs1, bs2)
}
\arguments{
\item{bs1}{a basisfd ...}

\item{bs2}{a basisfd ...}
}
\description{
Constructor for basis2Dfd objects
}
\keyword{internal}
