% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/basisClass.R
=======
% Please edit documentation in R/basisEmpClass.R
>>>>>>> 114fd85199cfe4746689a5a8c0913c653feb42fd
\docType{methods}
\name{basisEmp-constructor}
\alias{basisEmp-constructor}
\title{Constructor for basisEmp objects,}
\usage{
basisEmp(support, grids, B)
}
\arguments{
\item{support}{a numeric ...}

\item{grids}{a numeric vector  ...}

\item{B}{a numeric matrix ...}
}
\description{
Constructor for basisEmp objects,
}
\keyword{internal}
