% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basismfdClass.R
\docType{methods}
\name{basismfd-constructor}
\alias{basismfd-constructor}
\title{Constructor for basismfd objects}
\usage{
basismfd(B)
}
\arguments{
\item{B}{a list of basisEmp or basisfd}
}
\description{
Constructor for basismfd objects
}
\keyword{internal}
