% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdMethods.R
\name{bimfdplot}
\alias{bimfdplot}
\title{Bivariate plot for mvmfd objects}
\usage{
bimfdplot(mvmfd_obj, type = "l", lty = 1, xlab = "", ylab = "", main = "", ...)
}
\arguments{
\item{mvmfd_obj}{An mvmfd object}

\item{type}{Type of plot ('l' for lines, 'p' for points, etc.)}

\item{lty}{Line type}

\item{xlab}{Label for the x-axis}

\item{ylab}{Label for the y-axis}

\item{main}{Main title}

\item{...}{Additional arguments for the matplot function}
}
\description{
Bivariate plot for mvmfd objects
}
\seealso{
\code{\link{mvmfd}}
}
