% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basismfdMethods.R
\name{dimSupp,basismfd-method}
\alias{dimSupp,basismfd-method}
\title{dimSupp for basismfd objects}
\usage{
\S4method{dimSupp}{basismfd}(object)
}
\description{
dimSupp for basismfd objects
}
\keyword{internal}
