% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basismfdMethods.R
\name{dimSupp}
\alias{dimSupp}
\title{Support dimension of functional data}
\usage{
dimSupp(object)
}
\arguments{
\item{object}{An object of  class \code{mfd}.}
}
\value{
a numeric vector, giving the support dimension of each variable.
}
\description{
This function returns the support dimension of an object of class  \code{basismfd}.
}
\examples{
x <- 1

}
