% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{dimSupp1}
\alias{dimSupp1}
\title{Support dimension of functional data}
\usage{
dimSupp1(object)
}
\arguments{
\item{object}{An object of  class \code{mfd}.
@return a numeric vector, giving the support dimension of each variable.}
}
\description{
This function returns the support dimension of an object of class  \code{mfd}.
}
