% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{-.mfd}
\alias{-.mfd}
\title{Subtract two 'mfd' objects}
\usage{
\method{-}{mfd}(obj1, obj2 = NULL)
}
\arguments{
\item{obj1}{An 'mfd' object}

\item{obj2}{An 'mfd' object or a scalar}
}
\value{
The difference between the two 'mfd' objects
}
\description{
Subtract two 'mfd' objects
}
