% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_approach.R
\name{eigen_approach}
\alias{eigen_approach}
\title{Eigen Approach}
\usage{
eigen_approach(mvmfd_obj, alpha, n, centerfns, penalty_type)
}
\arguments{
\item{mvmfd_obj}{The mvmfd object.}

\item{alpha}{The alpha parameter.}

\item{n}{The number of iterations.}

\item{centerfns}{Logical indicating whether to center functions.}

\item{penalty_type}{The type of penalty.}
}
\value{
A list containing the results.
}
\description{
Perform eigen approach for mvmfd_obj.
}
