% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectricalPowerDataset.R
\docType{data}
\name{electrical_power_data}
\alias{electrical_power_data}
\title{Electrical Power Dataset}
\format{
A bivariate functional data object of class `mvmfd` with the following fields:
\describe{
  \item{voltage}{Voltage measurements in volts.}
  \item{power}{Calculated power consumption in watts.}
}
}
\source{
The dataset was collected from https://weather.com.
}
\description{
This dataset contains measurements of voltage and electrical power consumption
recorded from a household between December 2006 and November 2010.
}
\examples{
\dontrun{
# Load the Electrical Power Dataset
data("electrical_power_data")
head(electrical_power_data)
}
}
