% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basismfdMethods.R
\name{eval.basismfd,list,basismfd-method}
\alias{eval.basismfd,list,basismfd-method}
\title{evalvaluation of basismfd objects}
\usage{
\S4method{eval.basismfd}{list,basismfd}(evalarg, basisobj)
}
\description{
evalvaluation of basismfd objects
}
\keyword{internal}
