% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basismfdClass.R
\name{eval_basismf_validity_check}
\alias{eval_basismf_validity_check}
\title{Check the validity of the evaluation}
\usage{
eval_basismf_validity_check(evalarg, dimSupp)
}
\arguments{
\item{evalarg}{A list or numeric vector}

\item{dimSupp}{Dimension of the `basismfd` object}
}
\description{
Check the validity of the evaluation
}
