% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basismfdMethods.R
\name{evalbasismfd,numeric,basismfd-method}
\alias{evalbasismfd,numeric,basismfd-method}
\title{evalbasismfd of basismfd objects}
\usage{
\S4method{evalbasismfd}{numeric,basismfd}(evalarg, basisobj)
}
\description{
evalbasismfd of basismfd objects
}
\keyword{internal}
