% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basismfdMethods.R
\name{evalbasismfd}
\alias{evalbasismfd}
\title{Values of basismfd Basis Functions}
\usage{
evalbasismfd(evalarg, basisobj)
}
\arguments{
\item{evalarg}{a list of numeric vectors of argument values at which the \code{basismfd} is to be evaluated.}

\item{basisobj}{an object of class \code{basismfd} defining basis functions whose values are to be computed.}
}
\value{
a matrix of basis function values with rows corresponding to argument values and columns to basis functions.
}
\description{
Computes the basis matrix evaluated at arguments in EVALARG associated with basismfd object
}
\examples{
x <- 1

}
