% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basismfdClass.R
\name{init_basismfd_check}
\alias{init_basismfd_check}
\title{Check the validity of the initializer}
\usage{
init_basismfd_check(basis)
}
\arguments{
\item{basis}{A list of `basisfd` objects}
}
\description{
Check the validity of the initializer
}
