% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{inprod_mfd}
\alias{inprod_mfd}
\title{Compute the inner product between two 'mfd' objects}
\usage{
inprod_mfd(mfd_obj1, mfd_obj2)
}
\arguments{
\item{mfd_obj1}{An 'mfd' object}

\item{mfd_obj2}{An 'mfd' object}
}
\value{
The inner product between the two 'mfd' objects
}
\description{
Compute the inner product between two 'mfd' objects
}
\seealso{
\code{\link{basismfd}}, \code{\link{mfd}}
}
