% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdMethods.R
\name{inprod_mvmfd}
\alias{inprod_mvmfd}
\title{Inner product of two mvmfd objects}
\usage{
inprod_mvmfd(mvmfd_obj1, mvmfd_obj2)
}
\arguments{
\item{mvmfd_obj1}{An mvmfd object}

\item{mvmfd_obj2}{An mvmfd object}
}
\value{
A matrix of inner products
}
\description{
Inner product of two mvmfd objects
}
\seealso{
\code{\link{mvmfd}}
}
