% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.basismfd}
\alias{is.basismfd}
\title{Check if an object is of class 'basismfd'}
\usage{
is.basismfd(fdobj)
}
\arguments{
\item{fdobj}{The object to check.}
}
\value{
TRUE if the object is of class 'basismfd', FALSE otherwise.
}
\description{
Check if an object is of class 'basismfd'
}
\seealso{
\code{\link{is.mvbasismfd}}, \code{\link{is.mfd}}, \code{\link{is.mvmfd}}
}
