% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mfd}
\alias{is.mfd}
\title{Check if an object is of class 'mfd'}
\usage{
is.mfd(fdobj)
}
\arguments{
\item{fdobj}{The object to check.}
}
\value{
TRUE if the object is of class 'mfd', FALSE otherwise.
}
\description{
Check if an object is of class 'mfd'
}
\seealso{
\code{\link{is.mvbasismfd}}, \code{\link{is.basismfd}}, \code{\link{is.mvmfd}}
}
