% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mvbasismfd}
\alias{is.mvbasismfd}
\title{Check if an object is of class 'mvbasismfd'}
\usage{
is.mvbasismfd(fdobj)
}
\arguments{
\item{fdobj}{The object to check.}
}
\value{
TRUE if the object is of class 'mvbasismfd', FALSE otherwise.
}
\description{
Check if an object is of class 'mvbasismfd'
}
\seealso{
\code{\link{is.basismfd}}, \code{\link{is.mfd}}, \code{\link{is.mvmfd}}
}
