% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mvmfd}
\alias{is.mvmfd}
\title{Check if an object is of class 'mvmfd'}
\usage{
is.mvmfd(fdobj)
}
\arguments{
\item{fdobj}{The object to check.}
}
\value{
TRUE if the object is of class 'mvmfd', FALSE otherwise.
}
\description{
Check if an object is of class 'mvmfd'
}
\seealso{
\code{\link{is.mvbasismfd}}, \code{\link{is.mfd}}, \code{\link{is.basismfd}}
}
