% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{length.mfd}
\alias{length.mfd}
\title{Length of an 'mfd' object}
\usage{
\method{length}{mfd}(mfd_obj)
}
\arguments{
\item{mfd_obj}{An 'mfd' object}
}
\value{
The number of observations in the 'mfd' object
}
\description{
Length of an 'mfd' object
}
