% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{length}
\alias{length}
\title{Length of an object of classes `mfd`or `mvmfd`.}
\usage{
length(x, ...)
}
\arguments{
\item{x}{An object of classes `mfd` or `mvmfd`.}

\item{...}{all `length` function arguments.}
}
\description{
Length of an object of an object of classes `mfd` or `mvmfd`.
}
