% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{mean.mfd}
\alias{mean.mfd}
\title{Compute the mean of an 'mfd' object}
\usage{
\method{mean}{mfd}(mfd_obj)
}
\arguments{
\item{mfd_obj}{An 'mfd' object}
}
\value{
An 'mfd' object representing the mean
}
\description{
Compute the mean of an 'mfd' object
}
\seealso{
\code{\link{basismfd}}, \code{\link{mfd}}
}
