% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdMethods.R
\name{mean_mvmfd}
\alias{mean_mvmfd}
\title{Mean of each variable in an mvmfd object}
\usage{
mean_mvmfd(mvmfd_obj)
}
\arguments{
\item{mvmfd_obj}{An mvmfd object}
}
\value{
An mvmfd object
}
\description{
Mean of each variable in an mvmfd object
}
\seealso{
\code{\link{mvbasismfd}}, \code{\link{mvmfd}}
}
