% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdClass.R
\name{mfd}
\alias{mfd}
\title{A Class of Multidimensional Functional Data objects}
\description{
The `mfd` class represents functional data ...
}
\examples{
x <- 1

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{basis}}{NULL,}

\item{\code{coefs}}{record vetorized of the coefs}

\item{\code{nobs}}{description}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mfd-new}{\code{mfd$new()}}
\item \href{#method-mfd-eval}{\code{mfd$eval()}}
\item \href{#method-mfd-print}{\code{mfd$print()}}
\item \href{#method-mfd-clone}{\code{mfd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfd-new"></a>}}
\if{latex}{\out{\hypertarget{method-mfd-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfd$new(argval = NULL, X, mdbs, method = "data")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argval}}{description}

\item{\code{X}}{description}

\item{\code{mdbs}}{a basismfd object}

\item{\code{method}}{description}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfd-eval"></a>}}
\if{latex}{\out{\hypertarget{method-mfd-eval}{}}}
\subsection{Method \code{eval()}}{
evalmfd
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfd$eval(evalarg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{evalarg}}{a list of numeric vector of argument values at which the \code{mfd} is to be evaluated.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix of evaluated values
Print method for `mfd` objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfd-print"></a>}}
\if{latex}{\out{\hypertarget{method-mfd-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfd$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to be passed to `print`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mfd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
