% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdClass.R
\name{mvfd}
\alias{mvfd}
\title{A Class of Multidimensional Functional Data objects}
\description{
The \code{mvfd} class represents functional data ...
}
\examples{
x <- 1

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coeff}}{a matrix with nrow=subjects and ncol=total number of basis ...}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{basis}}{a basismfd object}

\item{\code{coeff}}{a matrix with nrow=subjects and ncol=total number of basis ...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mvfd-new}{\code{mvfd$new()}}
\item \href{#method-mvfd-eval}{\code{mvfd$eval()}}
\item \href{#method-mvfd-clone}{\code{mvfd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvfd-new"></a>}}
\if{latex}{\out{\hypertarget{method-mvfd-new}{}}}
\subsection{Method \code{new()}}{
Constructor for mvfd objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvfd$new(mfd_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mfd_list}}{a list of mfd objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvfd-eval"></a>}}
\if{latex}{\out{\hypertarget{method-mvfd-eval}{}}}
\subsection{Method \code{eval()}}{
evalmfd
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvfd$eval(evalarg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{evalarg}}{a list of numeric vector of argument values at which the \code{mvfd} is to be evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvfd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mvfd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvfd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
