% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{nObs}
\alias{nObs}
\title{Get the number of observations}
\usage{
nObs(object)
}
\arguments{
\item{object}{An object of class \code{mfd}}
}
\value{
The number of observations in \code{mfd}.
}
\description{
This functions returns the number of observations in a \code{mfd} object.
}
