% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{norm_mfd}
\alias{norm_mfd}
\title{Compute the norm of an object of class `mfd`}
\usage{
norm_mfd(mfd_obj)
}
\arguments{
\item{mfd_obj}{An object of class `mfd`}
}
\value{
The norm vector of the an object of class `mfd`
}
\description{
Compute the norm of an object of class `mfd`
}
\seealso{
\code{\link{basismfd}}, \code{\link{mfd}}
}
