% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdMethods.R
\name{norm_mvmfd}
\alias{norm_mvmfd}
\title{Compute the norm of an object of class `mvmfd`}
\usage{
norm_mvmfd(mvmfd_obj)
}
\arguments{
\item{mvmfd_obj}{An `mvmfd` object}
}
\value{
The norm vector of the an object of class `mvmfd`
}
\description{
Compute the norm of an object of class `mvmfd`
}
\seealso{
\code{\link{mvmfd}},\code{\link{mvbasismfd}}
}
