% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{nVar}
\alias{nVar}
\title{Get the number of variables}
\usage{
nVar(object)
}
\arguments{
\item{object}{An object of class \code{mfd}}
}
\value{
The number of variables in \code{mfd}.
}
\description{
This functions returns the number of variables in a \code{mfd} object.
}
