% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_fun.R
\name{pen_fun}
\alias{pen_fun}
\title{Penalty Function}
\usage{
pen_fun(data, devorder = 2, type)
}
\arguments{
\item{data}{The data object.}

\item{devorder}{The order of the derivative.}

\item{type}{The type of penalty.}
}
\value{
The penalty matrix.
}
\description{
Calculate the penalty matrix for data.
}
