% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdMethods.R
\name{plot.mvmfd}
\alias{plot.mvmfd}
\title{Plotting method for mvmfd objects}
\usage{
\method{plot}{mvmfd}(mvmfd_obj, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{mvmfd_obj}{An mvmfd object}

\item{xlab}{Label for the x-axis}

\item{ylab}{Label for the y-axis}

\item{...}{Additional arguments for the plot function}
}
\description{
Plotting method for mvmfd objects
}
\seealso{
\code{\link{mvmfd}}
}
