% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{plot}
\alias{plot}
\title{plots an object of classes `mfd`, `mvmfd` or `remfpca`}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{An object of classes `mfd`, `mvmfd` or `remfpca`}

\item{...}{all `plot` function arguments.}
}
\description{
plot an object of classes `mfd`, `mvmfd` or `remfpca`
}
