% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{plot_mfd}
\alias{plot_mfd}
\title{plot an `mfd` object}
\usage{
plot_mfd(
  mfd_obj,
  obs = 1,
  xlab = "",
  ylab = "",
  main = "",
  type = "l",
  lty = 1,
  ...
)
}
\arguments{
\item{mfd_obj}{An `mfd` object}

\item{obs}{Observation number to plot (default: 1)}

\item{xlab}{x-axis label (default: "")}

\item{ylab}{y-axis label (default: "")}

\item{main}{Title of the plot (default: "")}

\item{type}{Type of plot (default: "l")}

\item{lty}{Line type (default: 1)}

\item{...}{Additional arguments passed to the plot function}
}
\description{
plot an `mfd` object
}
\seealso{
\code{\link{basismfd}}, \code{\link{mfd}}
}
