% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdMethods.R
\name{plot_mvmfd}
\alias{plot_mvmfd}
\title{Plotting `mvmfd` objects}
\usage{
plot_mvmfd(mvmfd_obj, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{mvmfd_obj}{An mvmfd object}

\item{xlab}{Label for the x-axis}

\item{ylab}{Label for the y-axis}

\item{...}{Additional arguments for the plot function}
}
\description{
Plotting  `mvmfd` objects
}
\seealso{
\code{\link{mvbasismfd}}, \code{\link{mvmfd}}
}
