% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remfpcaMethods.R
\name{plot_remfpca}
\alias{plot_remfpca}
\title{Plotting Method for remfpca Objects}
\usage{
plot_remfpca(
  remfpca_obj,
  comp_index = NULL,
  var_index = NULL,
  ask = TRUE,
  expand = NULL,
  nx = 100,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{remfpca_obj}{An object of class `remfpca` representing the results of reduced-rank functional principal component analysis.}

\item{comp_index}{Numeric vector specifying the indices of the principal components to plot.
If NULL, all components will be plotted.}

\item{var_index}{Numeric vector specifying the indices of the variables to plot.
If NULL, all variables will be plotted.}

\item{ask}{Logical indicating whether to prompt the user before each plot.}

\item{expand}{Numeric vector specifying the expansion factor for the width of the confidence bands.
If NULL, a default value will be used based on the eigenvalues.}

\item{nx}{Integer specifying the number of points to generate along the x-axis.}

\item{xlab}{Character vector specifying the labels for the x-axis.}

\item{ylab}{Character vector specifying the labels for the y-axis.}

\item{...}{Additional arguments to be passed to the plot function.}
}
\value{
None
}
\description{
This function generates plots for remfpca objects, showing the functional principal components
and the corresponding variability for each variable.
}
