% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdMethods.R
\name{+.mvmfd}
\alias{+.mvmfd}
\title{Addition of two mvmfd objects}
\usage{
\method{+}{mvmfd}(obj1, obj2 = NULL)
}
\arguments{
\item{obj1}{An mvmfd object}

\item{obj2}{An optional mvmfd object}
}
\value{
An mvmfd object
}
\description{
Addition of two mvmfd objects
}
