% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{print.mfd}
\alias{print.mfd}
\title{A print method for univariate functional data}
\usage{
\method{print}{mfd}(object, ...)
}
\description{
This function prints basic information about a \code{mfd} object. This is
the standard console output for \code{mfd} objects.
}
