% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refpcaClass.R
\name{remfpca}
\alias{remfpca}
\title{A Class for ReMFPCA objects}
\description{
The \code{remfpca} class represents functional data ...
}
\examples{
x <- 1

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{basis}}{a basismfd object}

\item{\code{coeff}}{a matrix with nrow=subjects and ncol=total number of basis ...}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{basis}}{a basismfd object}

\item{\code{coeff}}{a matrix with nrow=subjects and ncol=total number of basis ...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-remfpca-new}{\code{remfpca$new()}}
\item \href{#method-remfpca-clone}{\code{remfpca$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remfpca-new"></a>}}
\if{latex}{\out{\hypertarget{method-remfpca-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{remfpca$new(
  mvmfd_obj,
  method = "eigen",
  ncomp,
  alpha = NULL,
  centerfns = TRUE,
  alpha_orth = TRUE,
  lambda_type = "variable",
  penalty_type = "coefpen"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remfpca-clone"></a>}}
\if{latex}{\out{\hypertarget{method-remfpca-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{remfpca$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
