% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{sd.mfd}
\alias{sd.mfd}
\title{Compute the standard deviation of an 'mfd' object}
\usage{
sd.mfd(mfd_obj)
}
\arguments{
\item{mfd_obj}{An 'mfd' object}
}
\value{
An 'mfd' object representing the standard deviation
}
\description{
Compute the standard deviation of an 'mfd' object
}
