% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{sd}
\alias{sd}
\title{sd of an object of classes `mfa`.}
\usage{
sd(x, ...)
}
\arguments{
\item{x}{An object of classes `mfd`, `mvmfd`.}

\item{...}{all `sd` function arguments.}
}
\description{
sdof an object of classes `mfd`.
}
