% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdarithmatic.R
\name{[.mfd}
\alias{[.mfd}
\title{Indexing into Multivariate Functional Data object}
\usage{
\method{[}{mfd}(Y, i = "index")
}
\arguments{
\item{Y}{An object of class \code{\link{mfd}}.}

\item{i}{The index value.}
}
\value{
An object of class \code{\link{mfd}}.
}
\description{
An indexing method for multivariate dunctional data  (\code{\link{mfd}}).
}
\examples{
\dontrun{
require(Rfssa)
load_github_data("https://github.com/haghbinh/Rfssa/blob/master/data/Callcenter.RData")
D <- matrix(sqrt(Callcenter$calls), nrow = 240)
u <- seq(0, 1, length.out = 240) # Define domain of functional data
d <- 22 # number of basis elements
Y <- mfd(list(D), list(list(d, "bspline")), list(u))
plot(Y)
plot(Y[10:15])
}

}
\seealso{
\code{\link{mfd}}
}
