% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{[.mvmfd}
\alias{[.mvmfd}
\title{Extract subsets of an 'mvmfd' object}
\usage{
\method{[}{mvmfd}(mvmfd_obj, i = "index", j = "index")
}
\arguments{
\item{mvmfd_obj}{An 'mvmfd' object}

\item{i}{An index or indices specifying the subsets to extract for the first dimension}

\item{j}{An index or indices specifying the subsets to extract for the second dimension}
}
\value{
An 'mfd' object containing the specified subsets
}
\description{
Extract subsets of an 'mvmfd' object
}
