% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdarithmatic.R
\name{*.mfd}
\alias{*.mfd}
\title{Multiplication of Multivariate Functional Data object}
\usage{
\method{*}{mfd}(Y1, Y2)
}
\arguments{
\item{Y1}{An object of class \code{\link{mfd}} or numeric.}

\item{Y2}{An object of class \code{\link{mfd}} or numeric.}
}
\value{
An object of class \code{\link{mfd}}.
}
\description{
A method for multivariate dunctional data  (\code{\link{mfd}}) pointwise multiplication and mfd-scalar multiplication. Note that if the mfd is multivariate
then a vector of numerics may be provided allowing for multiplication of different variables by different scalars. For example, multivariate mfd-numeric multiplication
follows the form of \code{Y*c(1,2)} if \code{Y} is a bivariate mfd.
}
\examples{
\dontrun{
require(Rfssa)
load_github_data("https://github.com/haghbinh/Rfssa/blob/master/data/Callcenter.RData")
D <- matrix(sqrt(Callcenter$calls), nrow = 240)
u <- seq(0, 1, length.out = 240) # Define domain of functional data
d <- 22 # number of basis elements
Y <- mfd(list(D), list(list(d, "bspline")), list(u))
plot(Y)
Ytimes <- Y * Y # multiply the multivariate dunctional data  by itself
plot(Ytimes)
Ytimes2 <- Y * 2 # multiply every term in the mfd by 2
plot(Ytimes2)
}

}
\seealso{
\code{\link{mfd}}
}
