% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdMethods.R
\name{*.mvmfd}
\alias{*.mvmfd}
\title{Multiplication of an `mvmfd` object with a scalar}
\usage{
\method{*}{mvmfd}(obj1, obj2)
}
\arguments{
\item{obj1}{An `mvmfd` object or a scalar}

\item{obj2}{An `mvmfd` object or a scalar}
}
\value{
An `mvmfd` object
}
\description{
Multiplication of an `mvmfd` object with a scalar
}
\seealso{
\code{\link{mvmfd}}
}
