% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisClass.R
\docType{class}
\name{basisEmp-class}
\alias{basisEmp-class}
\alias{basisEmp}
\alias{basisEmp,matrix,list,matrix-method}
\title{A class for univariate Empirical functional data}
\usage{
\S4method{basisEmp}{matrix,list,matrix}(support, grids, B)
}
\arguments{
\item{support}{a 2*k matrix  where the first row specify min and second max and k is dimension...}

\item{grids}{a list of length k  ...}

\item{B}{a matrix ...}
}
\description{
A class for univariate Empirical functional data
}
\section{Methods (by generic)}{
\itemize{
\item \code{basisEmp}: Constructor for basisEmp objects when .....
}}

\section{Slots}{

\describe{
\item{\code{support}}{a matrix ...}

\item{\code{grids}}{a numeric ....}

\item{\code{B}}{a matrix ....}
}}

