% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisEmpClass.R
\docType{class}
\name{basisEmp-class}
\alias{basisEmp-class}
\alias{basisEmp}
\alias{basisEmp,matrix,list,matrix-method}
\alias{basisEmp,numeric,numeric,matrix-method}
\title{A class for univariate Empirical functional data}
\usage{
\S4method{basisEmp}{matrix,list,matrix}(support, grids, Basis)

\S4method{basisEmp}{numeric,numeric,matrix}(support, grids, Basis)
}
\arguments{
\item{support}{a numeric ...}

\item{grids}{a numeric ...}

\item{Basis}{a matrix ...}
}
\description{
A class for univariate Empirical functional data
}
\section{Methods (by generic)}{
\itemize{
\item \code{basisEmp}: Constructor for basisEmp objects when .....

\item \code{basisEmp}: Constructor for basisEmp objects .....
}}

\section{Slots}{

\describe{
\item{\code{support}}{a numeric ...}

\item{\code{grids}}{a numeric ....}

\item{\code{Basis}}{a matrix ....}
}}

