% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/basisClass.R
=======
% Please edit documentation in R/basisEmpClass.R
>>>>>>> 114fd85199cfe4746689a5a8c0913c653feb42fd
\docType{class}
\name{basisEmp-class}
\alias{basisEmp-class}
\alias{basisEmp}
<<<<<<< HEAD
\alias{basisEmp,matrix,list,matrix-method}
\title{A class for univariate Empirical functional data}
\usage{
\S4method{basisEmp}{matrix,list,matrix}(support, grids, B)
}
\arguments{
\item{support}{a 2*k matrix  where the first row specify min and second max and k is dimension...}

\item{grids}{a list of length k  ...}
=======
\alias{basisEmp,numeric,numeric,matrix-method}
\title{A class for univariate Empirical functional data}
\usage{
\S4method{basisEmp}{numeric,numeric,matrix}(support, grids, B)
}
\arguments{
\item{support}{a numeric ...}

\item{grids}{a numeric ...}
>>>>>>> 114fd85199cfe4746689a5a8c0913c653feb42fd

\item{B}{a matrix ...}
}
\description{
A class for univariate Empirical functional data
}
\section{Methods (by generic)}{
\itemize{
\item \code{basisEmp}: Constructor for basisEmp objects when .....
}}

\section{Slots}{

\describe{
<<<<<<< HEAD
\item{\code{support}}{a matrix ...}
=======
\item{\code{support}}{a numeric ...}
>>>>>>> 114fd85199cfe4746689a5a8c0913c653feb42fd

\item{\code{grids}}{a numeric ....}

\item{\code{B}}{a matrix ....}
}}

