% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisEmpClass.R
\docType{methods}
\name{basisEmp-constructor}
\alias{basisEmp-constructor}
\title{Constructor for basisEmp objects,}
\usage{
basisEmp(support, grids, Basis)
}
\arguments{
\item{support}{a numeric ...}

\item{grids}{a numeric vector  ...}

\item{Basis}{a numeric matrix ...}
}
\description{
Constructor for basisEmp objects,
}
\keyword{internal}
