% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basismfdClass.R
\docType{class}
\name{basismfd-class}
\alias{basismfd-class}
\alias{basismfd}
\alias{basismfd,list-method}
\alias{basismfd,basisEmp-method}
\alias{basismfd,basisfd-method}
\alias{basismfd,basis2Dfd-method}
\title{A Class of multivariate functional data objects}
\usage{
\S4method{basismfd}{list}(B)

\S4method{basismfd}{basisEmp}(B)

\S4method{basismfd}{basisfd}(B)

\S4method{basismfd}{basis2Dfd}(B)
}
\arguments{
\item{B}{a list ...}
}
\description{
The \code{basismfd} class represents functional data ...
}
\section{Methods (by generic)}{
\itemize{
\item \code{basismfd}: Constructor for basismfd objects when both \code{support} and \code{B} are given as list

\item \code{basismfd}: Constructor for basismfd objects when \code{support} is given as vector
and \code{B} is given as matrix.

\item \code{basismfd}: Constructor for basismfd objects nd \code{B} is given as basisfd(from fda package)

\item \code{basismfd}: Constructor for basismfd objects
}}

\section{Slots}{

\describe{
\item{\code{p}}{a numeric....}

\item{\code{support}}{a matrix ...}

\item{\code{B}}{a list ....}
}}

\examples{
require(fda)
require(ReMFPCA)
bs_fd <- create.fourier.basis(c(0,2*pi),5)
argval <- seq(0,2*pi,1)
b <- eval.basis(argval,bs_fd)
bs_Em <- basisEmp(c(0,2*pi),argval,b)
B <- list(bs_fd,bs_Em)

bs1 <- basismfd(B=bs_Em)
bs1@support

bs2 <- basismfd(bs_fd)
bs2@support
bs3 <- basismfd(B)
bs3@support
}
