% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisClass.R
\docType{class}
\name{basismfd-class}
\alias{basismfd-class}
\alias{basismfd}
\alias{basismfd,list-method}
\alias{basismfd,basisEmp-method}
\alias{basismfd,basisfd-method}
\title{A Class of multivariate functional data objects}
\usage{
\S4method{basismfd}{list}(B)

\S4method{basismfd}{basisEmp}(B)

\S4method{basismfd}{basisfd}(B)
}
\arguments{
\item{B}{a list ...}

\item{support}{a list ...}
}
\description{
The \code{basismfd} class represents functional data ...
}
\section{Methods (by generic)}{
\itemize{
\item \code{basismfd}: Constructor for basismfd objects when both \code{support} and \code{B} are given as list

\item \code{basismfd}: Constructor for basismfd objects when \code{support} is given as vector
and \code{B} is given as matrix.

\item \code{basismfd}: Constructor for basismfd objects nd \code{B} is given as basisfd(from fda package)
}}

\section{Slots}{

\describe{
\item{\code{p}}{a numeric....}

\item{\code{support}}{a matrix ...}

\item{\code{B}}{a list ....}
}}

