% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisClass.R
\docType{methods}
\name{basismfd-constructor}
\alias{basismfd-constructor}
\title{Constructor for basismfd objects, third argument (B) passed as matrix or array of numerics}
\usage{
basismfd(p, support, B)
}
\arguments{
\item{p}{a numeric specify number of variables}

\item{support}{a list ...}

\item{B}{a list ...}
}
\description{
Constructor for basismfd objects, third argument (B) passed as matrix or array of numerics
}
\keyword{internal}
