% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basismfdClass.R
\name{basismfd}
\alias{basismfd}
\title{A Class of Multidimensional Basis Functions}
\description{
The \code{basismfd} class represents functional data ...
}
\examples{
x <- 1

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{basis}}{a list of basisfd objects}

\item{\code{dimSupp}}{a positive integer specify the dimension...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-basismfd-new}{\code{basismfd$new()}}
\item \href{#method-basismfd-eval}{\code{basismfd$eval()}}
\item \href{#method-basismfd-clone}{\code{basismfd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basismfd-new"></a>}}
\if{latex}{\out{\hypertarget{method-basismfd-new}{}}}
\subsection{Method \code{new()}}{
Constructor for basismfd objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basismfd$new(basis)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{basis}}{a list of basisfd objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basismfd-eval"></a>}}
\if{latex}{\out{\hypertarget{method-basismfd-eval}{}}}
\subsection{Method \code{eval()}}{
evalbasismfd
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basismfd$eval(evalarg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{evalarg}}{a list of numeric vector of argument values at which the \code{basismfd} is to be evaluated.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basismfd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-basismfd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basismfd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
