% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdarithmatic.R
\name{-.mfd}
\alias{-.mfd}
\title{Subtraction of Multivariate Functional Data object}
\usage{
\method{-}{mfd}(Y1, Y2)
}
\arguments{
\item{Y1}{An object of class \code{\link{mfd}} or numeric.}

\item{Y2}{An object of class \code{\link{mfd}} or numeric.}
}
\value{
An object of class \code{\link{mfd}}.
}
\description{
A method for multivariate dunctional data  (\code{\link{mfd}}) subtraction and mfd-scalar subtraction. Note that if the mfd is multivariate
then a vector of numerics may be provided allowing for subtraction of different scalars from different variables. For example, multivariate mfd-numeric subtraction
follows the form of \code{Y-c(1,2)} if \code{Y} is a bivariate mfd.
}
\examples{
\dontrun{
require(Rfssa)
load_github_data("https://github.com/haghbinh/Rfssa/blob/master/data/Callcenter.RData")
D <- matrix(sqrt(Callcenter$calls), nrow = 240)
u <- seq(0, 1, length.out = 240) # Define domain of functional data
d <- 22 # number of basis elements
Y <- mfd(list(D), list(list(d, "bspline")), list(u))
plot(Y)
Yminus <- Y[1:4] - Y[5:8] # subtract the multivariate dunctional data  to itself
plot(Yminus)
Yminus2 <- Y - 2 # subtract 2 to every term in the multivariate dunctional data 
plot(Yminus2)
}

}
\seealso{
\code{\link{mfd}}
}
