% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdMethods.R
\name{inprod_mvmfd}
\alias{inprod_mvmfd}
\title{Compute the inner product between two objects of class `mvmfd`}
\usage{
inprod_mvmfd(mvmfd_obj1, mvmfd_obj2)
}
\arguments{
\item{mvmfd_obj1}{An `mvmfd` object}

\item{mvmfd_obj2}{An `mvmfd` object}
}
\value{
The inner products matrix between the two `mvmfd` objects
}
\description{
Compute the inner product between two objects of class `mvmfd`
}
\seealso{
\code{\link{mvmfd}},\code{\link{mvbasismfd}}
}
