% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{mean}
\alias{mean}
\title{mean of an object of classes `mfd`or `mvmfd`.}
\usage{
mean(x, ...)
}
\arguments{
\item{x}{An object of classes `mfd` or `mvmfd`.}

\item{...}{all `mean` function arguments.}
}
\value{
An object of class `mfd`
}
\description{
mean of an object of classes `mfd`or `mvmfd`.
}
