% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvbasismfdClass.R
\name{Mvbasismfd}
\alias{Mvbasismfd}
\title{Constructor for mvbasismfd objects}
\usage{
Mvbasismfd(basis)
}
\arguments{
\item{basis}{A list of basisfd objects}
}
\description{
Constructor for mvbasismfd objects
}
