% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvbasismfdClass.R
\name{mvbasismfd}
\alias{mvbasismfd}
\title{A Class of Multidimensional Functional Data objects}
\description{
The \code{basismfd} class represents functional data ...
}
\examples{
x <- 1

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{basis}}{a list of basisfd objects}

\item{\code{dimSupp}}{a positive integer specify the dimension...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mvbasismfd-new}{\code{mvbasismfd$new()}}
\item \href{#method-mvbasismfd-eval}{\code{mvbasismfd$eval()}}
\item \href{#method-mvbasismfd-clone}{\code{mvbasismfd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvbasismfd-new"></a>}}
\if{latex}{\out{\hypertarget{method-mvbasismfd-new}{}}}
\subsection{Method \code{new()}}{
Constructor for mvbasismfd objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvbasismfd$new(basis)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{basis}}{a list of basisfd objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvbasismfd-eval"></a>}}
\if{latex}{\out{\hypertarget{method-mvbasismfd-eval}{}}}
\subsection{Method \code{eval()}}{
evalbasismfd
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvbasismfd$eval(evalarg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{evalarg}}{a list of numeric vector of argument values at which the \code{basismfd} is to be evaluated.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvbasismfd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mvbasismfd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvbasismfd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
