% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdClass.R
\name{mvmfd}
\alias{mvmfd}
\title{A Class of Multidimensional Functional Data objects}
\description{
The \code{mvmfd} class represents functional data ...
}
\examples{
x <- 1

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coeff}}{a matrix with nrow=subjects and ncol=total number of basis ...}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{basis}}{a basismfd object}

\item{\code{coeff}}{a matrix with nrow=subjects and ncol=total number of basis ...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mvmfd-new}{\code{mvmfd$new()}}
\item \href{#method-mvmfd-eval}{\code{mvmfd$eval()}}
\item \href{#method-mvmfd-print}{\code{mvmfd$print()}}
\item \href{#method-mvmfd-clone}{\code{mvmfd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvmfd-new"></a>}}
\if{latex}{\out{\hypertarget{method-mvmfd-new}{}}}
\subsection{Method \code{new()}}{
Constructor for mvmfd objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvmfd$new(mfd_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mfd_list}}{a list of mfd objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvmfd-eval"></a>}}
\if{latex}{\out{\hypertarget{method-mvmfd-eval}{}}}
\subsection{Method \code{eval()}}{
evalmfd
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvmfd$eval(evalarg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{evalarg}}{a list of numeric vector of argument values at which the \code{mvmfd} is to be evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvmfd-print"></a>}}
\if{latex}{\out{\hypertarget{method-mvmfd-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvmfd$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvmfd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mvmfd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvmfd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
