% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmfdMethods.R
\name{norm_mvmfd}
\alias{norm_mvmfd}
\title{Norm of an mvmfd object}
\usage{
norm_mvmfd(mvmfd_obj)
}
\arguments{
\item{mvmfd_obj}{An mvmfd object}
}
\value{
The norm of the mvmfd object
}
\description{
Norm of an mvmfd object
}
