% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_fun.R
\name{pen_fun}
\alias{pen_fun}
\title{Penalty Function}
\usage{
pen_fun(data, devorder = 2, type)
}
\arguments{
\item{data}{an object of class `mvmfd`.}

\item{devorder}{The order of the derivative.}

\item{type}{The type of penalty. The types "coefpen" and "basispen" is supported.}
}
\value{
The penalty matrix.
}
\description{
Calculate the penalty matrix for `mvmfd` objects.
}
