% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refpcaClass.R
\name{Remfpca}
\alias{Remfpca}
\title{Create a ReMFPCA object}
\usage{
Remfpca(
  mvmfd_obj,
  method = "eigen",
  ncomp,
  alpha = NULL,
  centerfns = TRUE,
  alpha_orth = TRUE,
  lambda_type = "variable",
  penalty_type = "coefpen"
)
}
\arguments{
\item{mvmfd_obj}{An mvmfd object representing the multivariate functional data.}

\item{method}{The method to be used for ReMFPCA. Currently, only "eigen" is supported.}

\item{ncomp}{The number of functional principal components to retain.}

\item{alpha}{A list or vector specifying the regularization parameter(s) for each variable.
If NULL, the regularization parameter is estimated internally.}

\item{centerfns}{Logical indicating whether to center the functional data before analysis.}

\item{alpha_orth}{Logical indicating whether to perform orthogonalization of the regularization parameters.}

\item{lambda_type}{The type of penalty to be applied on the basis functions. Default is "variable".}

\item{penalty_type}{The type of penalty to be applied on the coefficients. Default is "coefpen".}
}
\value{
An object of class `remfpca`.
}
\description{
Create an instance of the `remfpca` class, which represents functional data and provides
methods for performing reduced-rank functional principal component analysis (ReMFPCA).
}
