% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{sd}
\alias{sd}
\title{Standard deviation of an object of class `mfd`.}
\usage{
sd(x, ...)
}
\arguments{
\item{x}{An object of class `mfd`}

\item{...}{all `sd` function arguments.}
}
\value{
An object of class `mfd`
}
\description{
Standard deviation an object of class `mfd`.
}
