% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{*.mfd}
\alias{*.mfd}
\title{Scalar multiplication of an `mfd` object}
\usage{
\method{*}{mfd}(obj1, obj2)
}
\arguments{
\item{obj1}{An `mfd` object or an scalar}

\item{obj2}{An `mfd` object or an scalar}
}
\value{
An `mfd` object
}
\description{
Scalar multiplication of an `mfd` object. One object must be an `mfd`, and the other one a scalar
}
\seealso{
\code{\link{basismfd}}, \code{\link{mfd}}
}
